修改ordergroup最大为255，mac10。15版本将增加std::filesystem，不再使用boost

+++ src/Utils/Utils.h
@@ -27,7 +27,7 @@
 template<> class _DLL_EXPORT XE::OrderGroup< INDEX > \
 { \
 public: \
-    static constexpr XE::uint64 group_value = INDEX; \
+    static constexpr XE::uint8 group_value = INDEX; \
 }; \
 typedef XE::OrderGroup< INDEX > NAME##Group;
